Various fixes for integration tests

+++ src/mender-update/update_module/v3/platform/c++17/update_module_call.cpp
@@ -145,12 +145,17 @@ void UpdateModule::StateRunner::ProcessFinishedHandler(State state, error::Error
 
 error::Error UpdateModule::AsyncSystemReboot(
 	events::EventLoop &event_loop, StateFinishedHandler handler) {
-	system_reboot_.reset(new SystemRebootRunner {vector<string> {"reboot"}, event_loop});
+	if (!system_reboot_) {
+		system_reboot_.reset(new SystemRebootRunner {vector<string> {"reboot"}, event_loop});
+	}
 
 	log::Info("Calling `reboot` command and waiting for system to restart.");
-	system_reboot_->proc.Start();
+	auto err = system_reboot_->proc.Start();
+	if (err != error::NoError) {
+		return err.WithContext("Unable to call system reboot command");
+	}
 
-	auto err = system_reboot_->proc.AsyncWait(event_loop, [](error::Error err) {
+	err = system_reboot_->proc.AsyncWait(event_loop, [](error::Error err) {
 		// Even if it returns, give the reboot ten minutes to kill us. `handler` will only
 		// be called from the timeout handler.
 		if (err != error::NoError) {
@@ -158,12 +163,13 @@ error::Error UpdateModule::AsyncSystemReboot(
 		}
 	});
 	if (err != error::NoError) {
-		return err.WithContext("Unable to call system reboot command");
+		return err.WithContext("Unable to wait for system reboot command");
 	}
 
 	system_reboot_->timeout.AsyncWait(chrono::minutes(10), [handler](error::Error err) {
 		if (err != error::NoError) {
 			handler(err.WithContext("UpdateModule::AsyncSystemReboot"));
+			return;
 		}
 
 		handler(error::Error(
